Complete Order And Test Api

+++ shop.Service/DTOs/OrderCommand/CancelOrderDto.cs
@@ -2,6 +2,6 @@
 {
     public class CancelOrderDto
     {
-        public long OrderId { get; set; }
+        public int OrderId { get; set; }
     }
 }
